requirements.txt
langchain>=0.2,<0.4
langchain-community>=0.2,<0.4
langchain-huggingface>=0.1
gradio>=4.0
huggingface_hub>=0.23
sentence-transformers>=2.2
faiss-cpu>=1.7
python-dotenv>=1.0
pdfminer.six>=20221105